+++ sql/queries/feed_follows.sql
@@ -19,6 +19,4 @@ INNER JOIN feeds ON feeds.id = feed_follows.feed_id
 WHERE feed_follows.user_id = $1;
 
 -- name: RemoveFeedFollow :exec
-DELETE FROM feed_follows
-WHERE feed_follows.user_id = $1
-AND feed_follows.feed_id = $2;
+DELETE FROM feed_follows WHERE user_id = $1 AND feed_id = $2;
